backend/requirements.txt
Flask
playwright
openpyxl
asyncio
psycopg2-binary
python-dotenv